crew_ai/requirements.txt
functions-framework==3.*
line-bot-sdk==3.16.0
crewai==0.22.5
langchain-google-genai==0.0.11
langchain-community==0.0.29
google-cloud-logging
python-dotenv==1.0.0
duckduckgo_search==5.1.0
